+++ CHANGELOG.md
@@ -7,6 +7,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+QOL improvements.
+
+### Added
+
+- A general progress bar to show the progress of achievements.
+- Fixed the UI on light-based themes.
+- Added a "SECRET" to see which achievements are and were hidden.
+
 ## Release 0.1.0
 
 Initial release.